[capture_triton] fix special kwargs path
I didn't test this path when creating the orchestrator. This PR fixes
that path to work in the capture_triton path. The problem is that we are
handling a value that is an int (in the capture_triton path) and a
ConstantVariable (in the Dynamo triton path) so we abstract that out in
the orchestrator.

Test Plan:
- new tests

ghstack-source-id: 468ab1840efd84b3b31d7174cf269b372134b5c3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/132143