[pipelining] Add schedule send/recv pass
Inserts send/recv ops where needed in a compute-only pipeline schedule.

Any F or B action will require a recv op for its input and a send op
for its output, except for at the ends of the pipeline.

To avoid hangs caused by mixed-up orderings of sends/recvs across ranks,
we pick one compute action at a time and insert both its send op (on
that rank's schedule), and the matching recv op for the recipient stage
(on the schedule for the rank for that stage).

Follow Up Work
- ignores batching (which is an optimization)
- ignores cases where a stage sends to another stage on the same rank,
  and should skip the send/recv and directly access memory

ghstack-source-id: 52faa81083e407c393766de4a778475e19311ad8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130378