[fx][proxy tensor] Don't set is_fx_tracing when tracing with make_fx
It's unlikely that users of make_fx want to use is_fx_tracing as a way to determine if make_fx is running; most users are probably actually checking for symbolic_trace. This disables the is_fx_tracing modifications for make_fx.

The specific use case here is for nested tensor, where we want to throw an assert to warn users of symbolic_trace that they need to wrap the NJT constructor, but we don't need to do this for make_fx.

ghstack-source-id: afdab200c9cdce4da47de9fde5413f3fd1cb0d95
Pull Request resolved: https://github.com/pytorch/pytorch/pull/134056